FerBot/backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic>=2.11.0
python-multipart>=0.0.17
openai>=1.68.0
PyPDF2>=3.0.1
pymupdf4llm>=0.0.17
python-dotenv>=1.1.0
httpx>=0.28.0
tiktoken>=0.8.0
numpy>=2.1.0
scikit-learn>=1.5.0
